% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataExplorer-package.R
\docType{package}
\name{DataExplorer-package}
\alias{DataExplorer}
\alias{DataExplorer-package}
\title{DataExplorer: histExplorer}
\description{
This explores histotical data in the borealis database.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/buskod21}
  \item Report bugs at \url{https://github.com/buskod21/issues}
}

}
\author{
\strong{Maintainer}: Busayo Kodaolu \email{buskod21@gmail.com}

}
\keyword{internal}
