% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{FattyacidMeta}
\alias{FattyacidMeta}
\title{Metadata for the Fattyacid data}
\format{
An object of class \code{data.frame} with 13 rows and 2 columns.
}
\usage{
FattyacidMeta
}
\description{
The "FattyacidMeta" dataset contains metadata and schema information
related to the "Fattyacid" data. It provides details about the structure and context
of the Fattyacid dataset.

For more comprehensive information about the "Fattyacid" dataset and the study,
refer to the borealis database
[here](https://borealisdata.ca/dataset.xhtml?persistentId=doi:10.5683/SP3/WVC09T).
}
\keyword{Fish}
\keyword{Milk}
\keyword{Monensin,}
\keyword{Omega-3}
\keyword{acids}
\keyword{composition,}
\keyword{fatty}
\keyword{oil,}
