% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Milk}
\alias{Milk}
\title{Example datasets}
\format{
An object of class \code{data.frame} with 48 rows and 10 columns.
}
\usage{
Milk
}
\description{
The dataset is derived from the borealis database
and is used to show how the DataExplorer App works
}
\keyword{datasets}
