% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{treeDistances}
\alias{treeDistances}
\alias{Fattyacid}
\title{Example datasets}
\format{
An object of class \code{data.frame} with 16 rows and 21 columns.
}
\usage{
Fattyacid
}
\description{
The dataset is derived from the borealis database
and is used to show how the DataExplorer App works
}
\keyword{datasets}
