% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{FattyacidMeta}
\alias{FattyacidMeta}
\title{Example datasets}
\format{
An object of class \code{data.frame} with 13 rows and 2 columns.
}
\usage{
FattyacidMeta
}
\description{
The dataset is derived from the borealis database
and is used to show how the DataExplorer App works
}
\keyword{datasets}
