% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{FattyacidMeta}
\alias{FattyacidMeta}
\title{Metadata for the Fattyacid data}
\format{
An object of class \code{data.frame} with 13 rows and 2 columns.
}
\usage{
FattyacidMeta
}
\description{
The FattyMeta contains metadata and schema for the Fattyacid data.
}
\details{
Example datasets

More information about found in the borealis database here
(https://borealisdata.ca/dataset.xhtml?persistentId=doi:10.5683/SP3/WVC09T)
}
\keyword{Fish}
\keyword{Milk}
\keyword{Monensin,}
\keyword{Omega-3}
\keyword{acids}
\keyword{composition,}
\keyword{fatty}
\keyword{oil,}
