% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Feed}
\alias{Feed}
\title{Example datasets}
\format{
An object of class \code{data.frame} with 112 rows and 13 columns.
}
\usage{
Feed
}
\description{
The dataset is derived from the borealis database
and is used to show how the DataExplorer App works
}
\keyword{datasets}
