% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Feed}
\alias{Feed}
\title{Cow Feed Intake Data}
\format{

}
\usage{
Feed
}
\description{
The "Feed" dataset is sourced from the borealis database and serves
as an illustrative dataset for showcasing the capabilities of the DataExplorer App.
}
\details{
This dataset provides insights into cow feed intake, including information
on treatment, fish oil concentration, monensin concentration, day of measurement,
concentrations of Dry Matter Intake (DMI) for concentrate, silage, and forage,
as well as nutritional metrics such as Crude Protein Intake (CPI), Neutral Detergent
Fibre Intake (NDFI), Ether Extract Intake (EEI), and Metabolizable Energy Intake (MEI).

More information about the dataset and the underlying study can be accessed in the
borealis database [here](https://borealisdata.ca/dataset.xhtml?persistentId=doi:10.5683/SP3/WVC09T).
}
\keyword{Fish}
\keyword{Milk}
\keyword{Monensin,}
\keyword{Omega-3}
\keyword{acids}
\keyword{composition,}
\keyword{fatty}
\keyword{oil,}
