% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{FeedMeta}
\alias{FeedMeta}
\title{Metadata for the Feed data}
\format{
An object of class \code{data.frame} with 19 rows and 2 columns.
}
\usage{
FeedMeta
}
\description{
The "FeedMeta" dataset serves as a metadata companion to the "Feed" dataset.
It contains essential metadata and schema information, offering insights into the structure
and context of the main "Feed" dataset.

For a comprehensive understanding of the "Feed" dataset and the underlying study,
refer to the borealis database
[here](https://borealisdata.ca/dataset.xhtml?persistentId=doi:10.5683/SP3/WVC09T).
}
\keyword{Fish}
\keyword{Milk}
\keyword{Monensin,}
\keyword{Omega-3}
\keyword{acids}
\keyword{composition,}
\keyword{fatty}
\keyword{oil,}
