% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Milk}
\alias{Milk}
\title{Cow Milk Production Data}
\format{

}
\usage{
Milk
}
\description{
The "Milk" dataset, derived from the borealis database, serves as a demonstration
of the functionalities of the DataExplorer App.
}
\details{
This dataset comprises 48 observations, each representing a distinct instance of
cow milk production. It includes essential information such as treatment details, fish oil
concentration (2%), monensin concentration (mg per kg-1 dietary dry matter), and the nutritional
composition of milk, measured in kilograms per day (kg d-1). Additionally, it provides the
percentage of fat, protein, and lactose in the milk.

For a comprehensive understanding of the dataset and the associated study, refer to the borealis
database [here](https://borealisdata.ca/dataset.xhtml?persistentId=doi:10.5683/SP3/WVC09T).
}
\keyword{Fish}
\keyword{Milk}
\keyword{Monensin,}
\keyword{Omega-3}
\keyword{acids}
\keyword{composition,}
\keyword{fatty}
\keyword{oil,}
